# contracts/openapi/runtime-controller-v1.yaml
openapi: 3.1.0
info:
  title: Antnest Runtime Controller Internal API
  version: 1.0.0
  description: Trusted-network lifecycle API. It performs no caller authentication.
servers:
  - url: http://runtime-controller:8080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200": {description: Process is alive}
  /readyz:
    get:
      operationId: ready
      responses:
        "200": {description: Service dependencies are ready}
  /internal/v1/runtimes/{agent_id}:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    put:
      operationId: prepareRuntime
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PrepareRuntime"}
      responses:
        "202": {$ref: "#/components/responses/AcceptedOperation"}
        "400": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
    get:
      operationId: getRuntime
      responses:
        "200":
          description: Runtime desired and observed state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Runtime"}
        "404": {$ref: "#/components/responses/Problem"}
    delete:
      operationId: retireRuntime
      description: Remove compute resources while retaining the workspace.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202": {$ref: "#/components/responses/AcceptedOperation"}
        "404": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/stop:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: stopRuntime
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202": {$ref: "#/components/responses/AcceptedOperation"}
  /internal/v1/runtimes/{agent_id}/purge:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: purgeRuntime
      description: Remove compute resources and the workspace.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202": {$ref: "#/components/responses/AcceptedOperation"}
  /internal/v1/runtimes/{agent_id}/network-policy:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    put:
      operationId: updateRuntimeNetworkPolicy
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                mode: {$ref: "#/components/schemas/NetworkMode"}
              additionalProperties: false
      responses:
        "202": {$ref: "#/components/responses/AcceptedOperation"}
  /internal/v1/runtimes/{agent_id}/work:begin:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: beginRuntimeWork
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/BeginWorkRequest"}
      responses:
        "200":
          description: Work admission outcome
          content:
            application/json:
              schema: {$ref: "#/components/schemas/BeginWorkResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/work:end:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: endRuntimeWork
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/EndWorkRequest"}
      responses:
        "200":
          description: Work closure result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EndWorkResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/process:exec:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: executeRuntimeProcess
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ExecProcessRequest"}
      responses:
        "200":
          description: Process execution outcome and captured output
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExecProcessResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/files:read:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: readRuntimeFile
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ReadFileRequest"}
      responses:
        "200":
          description: File read outcome and content
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ReadFileResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/files:write:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: writeRuntimeFile
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WriteFileRequest"}
      responses:
        "200":
          description: File write outcome
          content:
            application/json:
              schema: {$ref: "#/components/schemas/WriteFileResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/files:edit:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: editRuntimeFile
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/EditFileRequest"}
      responses:
        "200":
          description: Exact text replacement outcome
          content:
            application/json:
              schema: {$ref: "#/components/schemas/EditFileResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/files:list:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: listRuntimeFiles
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ListFilesRequest"}
      responses:
        "200":
          description: Directory listing outcome
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ListFilesResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtimes/{agent_id}/operations:cancel:
    parameters:
      - $ref: "#/components/parameters/AgentID"
    post:
      operationId: cancelRuntimeWorkOperation
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CancelWorkOperationRequest"}
      responses:
        "200":
          description: Cancellation outcome
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CancelWorkOperationResponse"}
        "400": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
        "409": {$ref: "#/components/responses/Problem"}
  /internal/v1/runtime-operations/{operation_id}:
    parameters:
      - name: operation_id
        in: path
        required: true
        schema: {type: string, minLength: 1}
    get:
      operationId: getRuntimeOperation
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Operation"}
        "404": {$ref: "#/components/responses/Problem"}
components:
  parameters:
    AgentID:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
        pattern: "^[A-Za-z0-9._-]+$"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 200}
  schemas:
    NetworkMode:
      type: string
      enum: [restricted, unrestricted]
    PrepareRuntime:
      type: object
      required: [image_ref, network_mode]
      properties:
        image_ref: {type: string, minLength: 1}
        network_mode: {$ref: "#/components/schemas/NetworkMode"}
      additionalProperties: false
    Runtime:
      type: object
      required:
        - agent_id
        - image_ref
        - network_mode
        - desired_state
        - status
        - desired_generation
        - observed_generation
        - network_policy_epoch
        - observed_policy_epoch
        - resource_version
        - updated_at
      properties:
        agent_id: {type: string}
        image_ref: {type: string}
        network_mode: {$ref: "#/components/schemas/NetworkMode"}
        desired_state: {type: string, enum: [active, stopped, retired, purged]}
        status: {type: string}
        desired_generation: {type: integer, minimum: 1}
        observed_generation: {type: integer, minimum: 0}
        network_policy_epoch: {type: integer, minimum: 1}
        observed_policy_epoch: {type: integer, minimum: 0}
        failure_code: {type: string}
        failure_detail: {type: string}
        resource_version: {type: integer, minimum: 1}
        updated_at: {type: string, format: date-time}
    Operation:
      type: object
      required: [operation_id, agent_id, kind, status, generation, updated_at]
      properties:
        operation_id: {type: string}
        agent_id: {type: string}
        kind: {type: string, enum: [prepare, stop, retire, purge, update_network]}
        status: {type: string, enum: [pending, running, succeeded, failed, unknown, superseded]}
        generation: {type: integer, minimum: 1}
        error_code: {type: string}
        error_detail: {type: string}
        updated_at: {type: string, format: date-time}
    EffectDisposition:
      type: string
      enum: [completed, not_started, unknown]
    EffectOutcome:
      type: object
      required: [disposition, reason]
      properties:
        disposition: {$ref: "#/components/schemas/EffectDisposition"}
        reason: {type: string, minLength: 1}
        message: {type: string, maxLength: 4096}
      additionalProperties: false
    WorkKind:
      type: string
      enum: [run, maintenance]
    WorkReference:
      type: object
      required: [work_id, work_epoch, work_session_id]
      properties:
        work_id: {type: string, minLength: 1, maxLength: 160}
        work_epoch:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        work_session_id: {type: string, minLength: 1, maxLength: 160}
    OperationReference:
      allOf:
        - $ref: "#/components/schemas/WorkReference"
        - type: object
          required: [operation_id, request_digest]
          properties:
            operation_id: {type: string, minLength: 1, maxLength: 160}
            request_digest:
              type: string
              pattern: "^sha256:[0-9a-f]{64}$"
    RootName:
      type: string
      enum: [workspace, system_skills]
    RootPath:
      type: object
      required: [root, path]
      properties:
        root: {$ref: "#/components/schemas/RootName"}
        path:
          type: string
          minLength: 1
          description: Path relative to the named root; absolute paths and parent traversal are rejected.
      additionalProperties: false
    EnvironmentVariable:
      type: object
      required: [name, value]
      properties:
        name: {type: string, minLength: 1}
        value: {type: string}
      additionalProperties: false
    BeginWorkRequest:
      allOf:
        - $ref: "#/components/schemas/WorkReference"
        - type: object
          required: [kind]
          properties:
            kind: {$ref: "#/components/schemas/WorkKind"}
    BeginWorkResponse:
      allOf:
        - $ref: "#/components/schemas/EffectOutcome"
        - type: object
          required: [accepted]
          properties:
            accepted: {type: boolean}
    EndWorkRequest:
      allOf:
        - $ref: "#/components/schemas/WorkReference"
    EndWorkResponse:
      type: object
      required: [closed, reason]
      properties:
        closed: {type: boolean}
        reason: {type: string, minLength: 1}
        message: {type: string, maxLength: 4096}
      additionalProperties: false
    ExecProcessRequest:
      allOf:
        - $ref: "#/components/schemas/OperationReference"
        - type: object
          required: [argv, working_dir, timeout_ms]
          properties:
            argv:
              type: array
              minItems: 1
              items: {type: string}
            working_dir: {$ref: "#/components/schemas/RootPath"}
            env:
              type: array
              items: {$ref: "#/components/schemas/EnvironmentVariable"}
            timeout_ms:
              type: integer
              format: uint64
              minimum: 1
              maximum: 86400000
              description: Process execution timeout in milliseconds.
    ExecProcessResponse:
      type: object
      required: [outcome, process_reaped, truncated, progress_dropped]
      properties:
        outcome: {$ref: "#/components/schemas/EffectOutcome"}
        exit_code: {type: integer, format: int32}
        stdout:
          type: string
          contentEncoding: base64
          description: Base64 encoding of the captured stdout bytes.
        stderr:
          type: string
          contentEncoding: base64
          description: Base64 encoding of the captured stderr bytes.
        process_reaped: {type: boolean}
        result_digest:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        truncated: {type: boolean}
        progress_dropped: {type: boolean}
      additionalProperties: false
    ReadFileRequest:
      allOf:
        - $ref: "#/components/schemas/OperationReference"
        - type: object
          required: [path, offset, limit]
          properties:
            path: {$ref: "#/components/schemas/RootPath"}
            offset: {type: integer, format: int64, minimum: 0}
            limit: {type: integer, format: int64, minimum: 1, maximum: 8388608}
    ReadFileResponse:
      type: object
      required: [outcome, truncated]
      properties:
        outcome: {$ref: "#/components/schemas/EffectOutcome"}
        content:
          type: string
          contentEncoding: base64
          description: Base64 encoding of the file content bytes.
        result_digest:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
        truncated: {type: boolean}
      additionalProperties: false
    WriteFileRequest:
      allOf:
        - $ref: "#/components/schemas/OperationReference"
        - type: object
          required: [path, content]
          properties:
            path: {$ref: "#/components/schemas/RootPath"}
            content:
              type: string
              contentEncoding: base64
              maxLength: 11184812
              description: Base64 encoding of at most 8 MiB of file content bytes.
    WriteFileResponse:
      type: object
      required: [outcome]
      properties:
        outcome: {$ref: "#/components/schemas/EffectOutcome"}
        bytes_written: {type: integer, format: int64, minimum: 0}
        result_digest:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
      additionalProperties: false
    EditFileRequest:
      allOf:
        - $ref: "#/components/schemas/OperationReference"
        - type: object
          required: [path, old_string, new_string]
          properties:
            path: {$ref: "#/components/schemas/RootPath"}
            old_string:
              type: string
              minLength: 1
              maxLength: 8388608
              description: Exact text that must occur exactly once in the existing file.
            new_string:
              type: string
              maxLength: 8388608
    EditFileResponse:
      type: object
      required: [outcome]
      properties:
        outcome: {$ref: "#/components/schemas/EffectOutcome"}
        bytes_written: {type: integer, format: int64, minimum: 0}
        result_digest:
          type: string
          pattern: "^sha256:[0-9a-f]{64}$"
      additionalProperties: false
    ListFilesRequest:
      allOf:
        - $ref: "#/components/schemas/OperationReference"
        - type: object
          required: [path, limit]
          properties:
            path: {$ref: "#/components/schemas/RootPath"}
            limit: {type: integer, minimum: 1, maximum: 4096}
    DirectoryEntry:
      type: object
      required: [name, is_dir, size]
      properties:
        name: {type: string, minLength: 1}
        is_dir: {type: boolean}
        size: {type: integer, format: int64, minimum: 0}
      additionalProperties: false
    ListFilesResponse:
      type: object
      required: [outcome, truncated]
      properties:
        outcome: {$ref: "#/components/schemas/EffectOutcome"}
        entries:
          type: array
          items: {$ref: "#/components/schemas/DirectoryEntry"}
        truncated: {type: boolean}
      additionalProperties: false
    CancelWorkOperationRequest:
      allOf:
        - $ref: "#/components/schemas/WorkReference"
        - type: object
          required: [operation_id]
          properties:
            operation_id: {type: string, minLength: 1, maxLength: 160}
    CancelWorkOperationResponse:
      type: object
      required: [outcome]
      properties:
        outcome: {$ref: "#/components/schemas/EffectOutcome"}
      additionalProperties: false
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        code: {type: string, minLength: 1}
        detail: {type: string}
        trace_id: {type: string, pattern: "^[0-9a-f]{32}$"}
      additionalProperties: false
  responses:
    AcceptedOperation:
      description: Lifecycle intent was durably accepted.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Operation"}
    Problem:
      description: Problem Details response
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}